main_window, web_view, progress_bar, create_web_view_finished moved to a common class
