+++ Project/MSVC2015/GUI/MediaConch_GUI.vcxproj.filters
@@ -92,22 +92,10 @@
     <ClCompile Include="..\..\..\Source\GUI\Qt\settingswindow.cpp">
       <Filter>Source Files\GUI</Filter>
     </ClCompile>
-    <ClCompile Include="_Automated\moc_checkerwindow.cpp">
-      <Filter>Source Files\_Automated</Filter>
-    </ClCompile>
-    <ClCompile Include="_Automated\moc_settingswindow.cpp">
-      <Filter>Source Files\_Automated</Filter>
-    </ClCompile>
-    <ClCompile Include="_Automated\moc_policieswindow.cpp">
-      <Filter>Source Files\_Automated</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\..\Source\GUI\Qt\helpwindow.cpp">
       <Filter>Source Files\GUI</Filter>
     </ClCompile>
     <ClCompile Include="_Automated\moc_helpwindow.cpp">
-      <Filter>Source Files\_Automated</Filter>
-    </ClCompile>
-    <ClCompile Include="_Automated\moc_displaywindow.cpp">
       <Filter>Source Files\_Automated</Filter>
     </ClCompile>
     <ClCompile Include="..\..\..\Source\GUI\Qt\progressbar.cpp">
@@ -251,11 +239,14 @@
     <ClCompile Include="..\..\..\Source\GUI\Qt\publicpolicieswindow.cpp">
       <Filter>Source Files\GUI</Filter>
     </ClCompile>
-    <ClCompile Include="_Automated\moc_publicpolicieswindow.cpp">
-      <Filter>Source Files\_Automated</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\..\Source\Common\PluginPreHook.cpp">
       <Filter>Source Files\Core</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\..\Source\GUI\Qt\commonwebwindow.cpp">
+      <Filter>Source Files\GUI</Filter>
+    </ClCompile>
+    <ClCompile Include="_Automated\moc_commonwebwindow.cpp">
+      <Filter>Source Files\_Automated</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -512,6 +503,9 @@
       <Filter>Header Files\Core</Filter>
     </ClInclude>
     <ClInclude Include="..\..\..\Source\GUI\Qt\publicpolicieswindow.h">
+      <Filter>Header Files\GUI</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\Source\GUI\Qt\commonwebwindow.h">
       <Filter>Header Files\GUI</Filter>
     </ClInclude>
   </ItemGroup>
