+++ Source/GUI/Qt/displaywindow.cpp
@@ -27,7 +27,7 @@ namespace MediaConch {
 // Constructor / Desructor
 //***************************************************************************
 
-DisplayWindow::DisplayWindow(MainWindow* m) : main_window(m), web_view(NULL), progress_bar(NULL), is_finished(false)
+DisplayWindow::DisplayWindow(MainWindow* m) : CommonWebWindow(m), is_finished(false)
 {
 }
 
@@ -55,7 +55,7 @@ void DisplayWindow::display_display()
     web_view->setPage(page);
 
     QObject::connect(web_view, SIGNAL(loadProgress(int)), progress_bar->get_progress_bar(), SLOT(setValue(int)));
-    QObject::connect(web_view, SIGNAL(loadFinished(bool)), this, SLOT(create_web_view_finished(bool)));
+    QObject::connect(web_view, SIGNAL(loadFinished(bool)), this, SLOT(on_loadFinished(bool)));
 
     QUrl url = QUrl("qrc:/html");
     if (!url.isValid())
