+++ Source/GUI/Qt/policieswindow.cpp
@@ -25,7 +25,7 @@ namespace MediaConch {
 // Constructor / Desructor
 //***************************************************************************
 
-PoliciesWindow::PoliciesWindow(MainWindow *parent) : main_window(parent), web_view(NULL), progress_bar(NULL)
+PoliciesWindow::PoliciesWindow(MainWindow *parent) : CommonWebWindow(parent)
 {
 }
 
@@ -222,7 +222,7 @@ void PoliciesWindow::create_html()
     web_view->setPage(page);
 
     QObject::connect(web_view, SIGNAL(loadProgress(int)), progress_bar->get_progress_bar(), SLOT(setValue(int)));
-    QObject::connect(web_view, SIGNAL(loadFinished(bool)), this, SLOT(create_web_view_finished(bool)));
+    QObject::connect(web_view, SIGNAL(loadFinished(bool)), this, SLOT(on_loadFinished(bool)));
 
     QUrl url = QUrl("qrc:/html");
     if (!url.isValid())
